// ClaimFlow/src/shared/components/forms/schemas/contact.schema.ts
import { z } from "zod";

export const contactSchema = z.object({
  name: z.string().min(2, "Name is too short"),
  email: z.string().email("Invalid email"),
  phone: z.string().optional(),
  subject: z.string().min(3, "Subject is too short"),
  message: z.string().min(10, "Message is too short"),
});

export type ContactFormValues = z.infer<typeof contactSchema>;